Create Rearrange-string-as-alternate-x-and-y-occurences
Hi Everyone,
This is the solution of GFG problem named "Rearrange string as alternate x and y occurences" & the problem stated as - 
Given a binary string s and two integers x and y. Arrange the given string in such a way so that '0' comes X-times then '1' comes Y-times and so on until one of the '0' or '1' is finished. Then concatenate the rest of the string and find the final string.

Example 1:

Input:
X = 1, Y = 1, S = "0011"
Output: "0101"
Explanation: we put 1 '0' and 1 '1'
alternatively.