Create Non-repetitiveString
Hello Everyone , this is the solution of GFG problem named "Non-repetitive String & the problem stated as -
Given a string S, the task is to check if a string is a non-repetitive or not.  A non-repetitive string is defined as a string that does not contain any different character between two same characters
.Input: S = "AABBCCCCC"
Output: 1
Explanation: No different character is 
placed between the two same characters.